chore: move events to libs/
